% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bypass}
\alias{sutter_habitat}
\alias{yolo_habitat}
\title{Bypass Habitat Area}
\format{
a 2 dimensional array [12 months, 21 years]
}
\usage{
sutter_habitat

yolo_habitat
}
\description{
The 1980-2000 suitable juvenile rearing habitat area in square meters
}
\details{
Calculated using the function \code{\link[=set_bypass_habitat]{set_bypass_habitat}}
}
\section{Sutter Bypass}{


The area within Sutter Bypass is the summation of the four sections listed below
\itemize{
  \item sutter1 = to Moulton Weir
  \item sutter2 = to Colusa Weir
  \item sutter3 = to Tisdale Weir
  \item sutter4 = below Tisdale Weir
}
}

\section{Yolo Bypass}{


The area within Yolo Bypass is the summation of the two sections listed below

\itemize{
  \item yolo1 = Fremont Weir to Sacramento Weir
  \item yolo2 = below Sacramento Weir
}
}

\keyword{datasets}
