% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{delta_habitat}
\alias{delta_habitat}
\title{Delta Habitat Array}
\format{
a 3 dimensional array [12 months, 21 years, 2 deltas]
}
\usage{
delta_habitat
}
\description{
The 1980-2000 delta rearing habitat
}
\details{
North Delta data is stored at [ , , 1] and South Delta data is stored at [ , , 2]
Values created using \code{\link[=delta_rearing_habitat]{delta_rearing_habitat}}
}
\keyword{datasets}
