% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{delta_prop_high_predation}
\alias{delta_prop_high_predation}
\title{Probability High Predation in the Deltas}
\format{
An object of class \code{numeric} of length 2.
}
\usage{
delta_prop_high_predation
}
\description{
Expert estimated high predation probabilities
}
\keyword{datasets}
