% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{estimated_probs}
\alias{estimated_probs}
\alias{prop_high_predation}
\alias{prob_strand_early}
\alias{prob_strand_late}
\alias{prob_nest_scoured}
\title{Expert Estimated Probabilities}
\format{
1 dimensional array [31 watersheds]
}
\source{
Various \href{https://s3-us-west-2.amazonaws.com/cvpia-reference-docs/AWP+Attachment+1.pdf}{CVPIA Science Integration Team: FY18 Decision Support Model activities and FY17 priorities Memorandum}

Compiled by: James T. Peterson \email{jt.peterson@oregonstate.edu}
}
\usage{
prop_high_predation

prob_strand_early

prob_strand_late

prob_nest_scoured
}
\description{
Experts estimated probabilities for predation, stranding, and nest scouring
}
\keyword{datasets}
