% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fry}
\alias{fr_fry}
\alias{wr_fry}
\alias{sr_fry}
\alias{st_fry}
\alias{lfr_fry}
\title{Fry Inchannel Habitat}
\format{
\itemize{
\item Steelhead: 3 dimensional array [31 watersheds, 12 months, 20 years]
\item Fall Run, Late Fall Run, Spring Run and Winter Run: 3 dimensional array [31 watersheds, 12 months, 21 years]
}
}
\usage{
fr_fry

wr_fry

sr_fry

st_fry

lfr_fry
}
\description{
The 1980-1999 fry suitable inchannel rearing habitat area in square meters for Steelhead
and 1979-2000 suitable inchannel rearing habitat area in square meters for Fall Run, Spring Run and Winter Run.
}
\details{
Values created using the function \code{\link[=set_instream_habitat]{set_instream_habitat}}
}
\keyword{datasets}
