% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gravel_size_to_prop_of_movement}
\alias{gravel_size_to_prop_of_movement}
\title{Gravel size to proportion of movement}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 8 rows and 4 columns.
}
\usage{
gravel_size_to_prop_of_movement
}
\description{
Gravel size to proportion of movement
}
\keyword{datasets}
