% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{habitat_array_to_dataframe}
\alias{habitat_array_to_dataframe}
\title{Habitat Array to Dataframes}
\usage{
habitat_array_to_dataframe(habitat_array)
}
\description{
converts the habitat array (e.g 'fr_spawn') and converts it a list
of dataframes. Each element in the list corresponds to a watershed habitat.
}
