% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{juvenile}
\alias{fr_juv}
\alias{wr_juv}
\alias{sr_juv}
\alias{st_juv}
\alias{lfr_juv}
\title{Juvenile Inchannel Habitat}
\format{
\itemize{
\item Steelhead: 3 dimensional array [31 watersheds, 12 months, 20 years]
\item Fall Run, Late Fall Run, Spring Run and Winter Run: 3 dimensional array [31 watersheds, 12 months, 21 years]
}
}
\usage{
fr_juv

wr_juv

sr_juv

st_juv

lfr_juv
}
\description{
The 1980-1999 juvenile suitable inchannel rearing habitat area in square meters for  
Steelhead and 1980-2000 juvenile suitable inchannel rearing habitat area in square meters
for Fall Run, Spring Run and Winter Run.
}
\details{
Values created using the function \code{\link[=set_instream_habitat]{set_instream_habitat}}
}
\keyword{datasets}
