% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-floodplain-habitat.R
\name{set_floodplain_habitat}
\alias{set_floodplain_habitat}
\title{Set Floodplain Habitat Area}
\usage{
set_floodplain_habitat(
  watershed,
  species,
  flow,
  flow2 = NULL,
  scenario_option = NULL
)
}
\arguments{
\item{watershed}{a watershed defined for the SIT model}

\item{species}{one of 'fr' (Fall Run), 'lfr' (Late Fall Run), 'sr' (Spring Run), or 'st' (Steelhead)}

\item{flow}{a flow value in cubic feet per second}
}
\value{
floodplain habitat value in square meters
}
\description{
This function returns an estimated floodplain area based on watershed, species and flow.
}
\section{Lower-mid Sacramento River}{

The Lower-mid Sacramento River has two nodes, one above Fremont Weir (C134) and one below (C160).
When calculating habitat for the Lower-Mid Sacramento river, calculate the habitat at
each flow node and sum them proportion to the length of stream above and below the weir:

35.6/58 * (habitat at C134) + 22.4/58 * (habitat at C160)


\strong{Regional Approximation:}
When a watershed has no associated floodplain modeling, an approximation is made.
[insert method description]
}

\examples{
# floodplain habitat value in square meters for Fall Run in the American River
set_floodplain_habitat("American River", "fr", 34652)

# floodplain habitat value in square meters for Steelhead in the American River
set_floodplain_habitat("American River", "st", 34652)
}
