% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spawn}
\alias{fr_spawn}
\alias{wr_spawn}
\alias{sr_spawn}
\alias{st_spawn}
\title{Spawning Habitat}
\format{
\itemize{
\item Steelhead: 3 dimensional array [31 watersheds, 12 months, 21 years]
\item Fall Run, Spring Run and Winter Run: 3 dimensional array [31 watersheds, 12 months, 22 years]
}
}
\usage{
fr_spawn

wr_spawn

sr_spawn

st_spawn
}
\description{
The 1979-1999 suitable spawning habitat area in square meters for Steelhead
and 1979-2000 suitable spawning habitat area in square meters for Fall Run, Spring Run and Winter Run.
}
\details{
Values created using the function \code{\link[=set_spawning_habitat]{set_spawning_habitat}}
}
\keyword{datasets}
