% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spawning_decay_multiplier}
\alias{spawning_decay_multiplier}
\title{Spawning Decay Multipliers}
\format{
array 31x22x12
}
\usage{
spawning_decay_multiplier
}
\description{
multiplier array to be used as decay per month and year for Spawning Habitat.
}
\details{
this decay multiplier is nested list with decays to represent differet runs, and different calsim flows.
}
\keyword{datasets}
