% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spawning_habitat_average}
\alias{spawning_habitat_average}
\title{Spawning Habitat Average}
\format{
List of 3 (fr, sr, wr) lenght 31 (watershed) named vectors
}
\usage{
spawning_habitat_average
}
\description{
habitat average by watershed in acres. This data is used to calcualte decays for all 
watersheds.
}
\keyword{datasets}
