% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{watershed_decay_status}
\alias{watershed_decay_status}
\title{Watershed Decay Status}
\format{
An object of class \code{logical} of length 31.
}
\usage{
watershed_decay_status
}
\description{
Watershed to TRUE/FALSE lookup indicating whether habitat decay occurs.
}
\keyword{datasets}
