% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{watershed_routing_indicators}
\alias{watershed_routing_indicators}
\alias{tisdale_bypass_watershed}
\alias{yolo_bypass_watershed}
\alias{south_delta_routed_watersheds}
\title{Watershed Routing}
\format{
1 dimensional array [31 watersheds]
}
\source{
Various \href{https://s3-us-west-2.amazonaws.com/cvpia-reference-docs/AWP+Attachment+1.pdf}{CVPIA Science Integration Team: FY18 Decision Support Model activities and FY17 priorities Memorandum}

Compiled by: James T. Peterson \email{jt.peterson@oregonstate.edu}
}
\usage{
tisdale_bypass_watershed

yolo_bypass_watershed

south_delta_routed_watersheds
}
\description{
Indicator of whether fish have access to a region
}
\keyword{datasets}
