% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{watershed_spawning_decays}
\alias{watershed_spawning_decays}
\title{Watershed Spawning Habitat Decays}
\format{
a list of 31 dataframes, each dataframe containing accumulated decay for the 20 year simulation
for each curve level ('min', 'avg', 'max') or 'none' for non-decay watersheds.
}
\usage{
watershed_spawning_decays
}
\description{
The total decay by watershed. This includes watersheds with no decay, these simply
have all values equal to 0. For watersheds with decay, there are three different curves,
'min', 'avg', 'max'.
}
\keyword{datasets}
