% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{weeks_flooded}
\alias{weeks_flooded}
\title{Floodplain Habitat Activation Duration}
\format{
a 3 dimensional array [31 watersheds, 12 months, 21 years]
}
\usage{
weeks_flooded
}
\description{
The 1980-2000 floodplain rearing habitat event duration in number of weeks
}
\details{
Values created using the function \code{\link[=get_weeks_flooded]{get_weeks_flooded}}
}
\keyword{datasets}
