% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spawning_decay_multiplier}
\alias{spawning_decay_multiplier}
\title{Spawning Decay Multiplier}
\format{
list of 31, 12x22 matrices.
}
\usage{
spawning_decay_multiplier
}
\description{
List of multiplier matrices to be used as decay per month and year for Spawning Habitat.
}
\keyword{datasets}
