% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{watershed_decays}
\alias{watershed_decays}
\title{Watershed Decays}
\format{
An object of class \code{list} of length 31.
}
\usage{
watershed_decays
}
\description{
The total decay by watershed. This includes watersheds with no decay, these simply
have all values equal to 0. For watersheds with decay, there are three different curves,
'min', 'avg', 'max'.
}
\keyword{datasets}
